FPGA implementation and modelsim simulation.
Modelsim simulation added with run.do and wave.do.
Also the project of quartus is ready with the wrapper and the module
that generates the 1HZ clock signal.
The MIPS processor of this project is able to execute only the
instructions add and addi, but it has the skeleton to add the other
instructions.

Andres Rivera Marquez.