Modified Datapath and working Modelsim Simulation
A multiplexer to allow j instruction is added to the datapath and the UC
has been modified to be able to execute the ori, j and beq instructions.

The testbench file was edited to show what instructions is being
executed by the MIPS processor using an enum type to display the
instruction in the waveform.

To-Do:
-Create the GPIO_i to allow the input from the switches.

Andres Rivera Marquez.