T7- Added Ori to read the GPIO_i
Added the instructions j to execute unconditional jumps in the
instruction code, and the instruction beq to create branches on equal
results of two registers. Also the ori instruction is added, but the
objective of this instruction is modified in order to have the ability
to read a GPIO_i port and save it in the RF.

Andres Rivera Marquez.